#1352 Reference data at Monitor: renaming of Page<Number> classes (no numbers anymore).
