+++ monitor/extensions/referencedata-wizards/src/main/java/org/datacleaner/monitor/server/wizard/stringpattern/regexswap/RegexSwapStringPatternReferenceDataWizardSession.java
@@ -45,7 +45,7 @@ final class RegexSwapStringPatternReferenceDataWizardSession extends AbstractRef
 
     @Override
     public WizardPageController firstPageController() {
-        return new RegexSwapStringPatternReferenceDataPage1(this);
+        return new RegexSwapStringPatternReferenceDataPageCategory(this);
     }
 
     @Override
