+++ monitor/extensions/referencedata-wizards/src/main/java/org/datacleaner/monitor/server/wizard/synonymcatalog/datastore/DatastoreSynonymCatalogReferenceDataWizardSession.java
@@ -44,7 +44,7 @@ final class DatastoreSynonymCatalogReferenceDataWizardSession extends DatastoreW
 
     @Override
     public WizardPageController firstPageController() {
-        return new DatastoreSynonymCatalogReferenceDataPage1(this);
+        return new DatastoreSynonymCatalogReferenceDataPageDatastore(this);
     }
 
     @Override
