+++ monitor/extensions/referencedata-wizards/src/main/resources/org/datacleaner/monitor/server/wizard/stringpattern/regexswap/RegexSwapStringPatternReferenceDataPageName.html
