+++ monitor/extensions/referencedata-wizards/src/main/resources/org/datacleaner/monitor/server/wizard/synonymcatalog/datastore/DatastoreSynonymCatalogReferenceDataPageSynonym.html
